Set log_*_consumer_threads to 1
Sync: monasca-persister: Change nr of threads for metrics from 4 to 1.
With 3 node cluster if we set log_*_consumer_threads to 3 all logs will
be processed on only one node.
As a workaround for now we change nr of this threads to 1 so that
processing logs will be balanced between all 3 node.
Affected components:
- monasca-log-metrics
- monasca-log-persister
- monasca-log-transformer
- monasca-persister

Change-Id: I1751555bf154a33ef300f342098b2d3a2c15fc25
Bug-Id: 13151