Do not install monasca-cli automatically
Following commit removes deploying of python-monascalcient
(a.k.a. monasca-cli role) from the CMM deployment.
It was recognized as the feature that was not fully
tested and should not be delivered at this point.

(cherry picked from commit 626dd20c485f3c851c4bdc9f8e9b5b2705cc9ac7)

Change-Id: Iadff43b6e0f822a1f1f0fddddb711e480d5e6165